Complete post API method for users table
